# docs/customizations/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mayan EDMS API",
    "description": "Free Open Source Electronic Document Management System",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v4"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "basePath": "/api/v4",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "paths": {
...
    "/document_types/{document_type_id}/ocr/settings/": {
      "get": {
        "operationId": "document_types_ocr_settings_read",
        "description": "Return the document type OCR settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "put": {
        "operationId": "document_types_ocr_settings_update",
        "description": "Set the document type OCR settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "patch": {
        "operationId": "document_types_ocr_settings_partial_update",
        "description": "Set the document type OCR settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "parameters": [
        {
          "name": "document_type_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
    "/document_types/{document_type_id}/parsing/settings/": {
      "get": {
        "operationId": "document_types_parsing_settings_read",
        "description": "Return the document type parsing settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "put": {
        "operationId": "document_types_parsing_settings_update",
        "description": "Set the document type parsing settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "patch": {
        "operationId": "document_types_parsing_settings_partial_update",
        "description": "Set the document type parsing settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "parameters": [
        {
          "name": "document_type_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/documents/{document_id}/tags/": {
      "get": {
        "operationId": "documents_tags_list",
        "description": "Returns a list of all the tags attached to a document.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "parameters": [
        {
          "name": "document_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/tags/": {
      "get": {
        "operationId": "tags_list",
        "description": "Returns a list of all the tags.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "tags_create",
        "description": "Create a new tag.",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "parameters": []
    },
    "/tags/{tag_id}/": {
      "get": {
        "operationId": "tags_read",
        "description": "Return the details of the selected tag.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "put": {
        "operationId": "tags_update",
        "description": "Edit the selected tag.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "patch": {
        "operationId": "tags_partial_update",
        "description": "Edit the selected tag.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "delete": {
        "operationId": "tags_delete",
        "description": "Delete the selected tag.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "parameters": [
        {
          "name": "tag_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
...
}